+++ Src/Plugs/Seeta/FaceSeeta.cpp
@@ -6,6 +6,7 @@
 #include "LandmarkerSeeta.h"
 #include "RecognizerSeeta.h"
 #include "FaceToolsSeeta.h"
+#include "FactoryFace.h"
 
 #include <QDir>
 
@@ -21,3 +22,17 @@ CFaceSeeta::CFaceSeeta(QObject *parent) : CFace(parent)
 
 CFaceSeeta::~CFaceSeeta()
 {}
+
+int CFaceSeeta::Initialize(CFactoryFace *pFactoryFace)
+{
+    int nRet = 0;
+    nRet = pFactoryFace->RegisterFace("Seeta", this, tr("Seeta"));
+    return nRet;
+}
+
+int CFaceSeeta::Clean(CFactoryFace* pFactoryFace)
+{
+    int nRet = 0;
+    nRet = pFactoryFace->RemoveFace("Seeta");
+    return nRet;
+}