+++ Src/Plugs/Seeta/FaceSeeta.h
@@ -1,0 +1,27 @@
+#ifndef CFACESEETA_H_KL_2019_10_21
+#define CFACESEETA_H_KL_2019_10_21
+
+#pragma once
+
+#include "Face.h"
+
+class CFaceSeeta : public CFace
+{
+    Q_OBJECT
+    Q_CLASSINFO("Author", "Kang Lin <kl222@126.com>")
+
+public:
+    CFaceSeeta(QObject* parent = nullptr);
+    virtual ~CFaceSeeta() override;
+    
+    Q_INTERFACES(CFace)
+#if (QT_VERSION >= QT_VERSION_CHECK(5, 0, 0))
+    Q_PLUGIN_METADATA(IID "KangLinStudio.Rabbit.FaceRecognizer.Plugs.Seeta.CFaceSeeta")
+#endif
+
+    int Initialize(CFactoryFace* pFactoryFace = nullptr) override;
+    int Clean(CFactoryFace* pFactoryFace = nullptr) override;
+
+};
+
+#endif // CFACESEETA_H_KL_2019_10_21