+++ Src/libfacedetection/FaceLibfacedetection.cpp
@@ -1,7 +1,0 @@
-#include "FaceLibfacedetection.h"
-#include "DetectorLibfacedetection.h"
-
-CFaceLibfacedetection::CFaceLibfacedetection(QObject *parent) : CFace(parent)
-{
-    m_pDetector = new CDetectorLibfacedetection(this);
-}