+++ ChangeLog.md
@@ -3,6 +3,7 @@
   + Modify Detect interface
   + Delete parameter interface
   + Add libfacedection 
+  + Use qt plugs manage ai libraries
 
 - v0.0.4
   + Add opencv face